serviceのcreate,read機能完成

+++ src/main/java/com/softbank/recipesitory/repository/RecipeRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.softbank.recipesitory.dao.RecipeDao;
 
-public interface RecipeRepo extends JpaRepository<RecipeDao, Long> {
+public interface RecipeRepo extends JpaRepository<RecipeDao, Integer> {
 
 }
